Fix transition from minigame to end scene
Immediately after a minigame ended, the board would be visible for a small amount of time before the fade transition.
This fix allows fading directly from the minigame to the end scene.
This doesn't screw up anything with the scene stack since `_pop_to_board` is called by `ScreenTransitions.fade_to_board` later on when the end scene is exited.